网站接口

+++ src/main/java/com/stk/website/comm/QueryPage.java
@@ -88,7 +88,7 @@ public class QueryPage implements Serializable {
 		this.targetPage = targetPage;
 	}
 
-	public int getLimitStart() {
+	public long getLimitStart() {
 		return (targetPage - 1) * pageSize;
 	}
